Make pythran validation Warning-free
Use -Wall -Werror for all tests. This spots a lot of minor issues:

* struct/class harmonization
* signed / unsigned comparison
* trailing characters in macro
* std::string::npos != -1
* functions missing the inline keywords
* replace the old  I::this_should_never_happen trick by a static assert

+++ pythran/pythonic/numpy/put.hpp
@@ -16,7 +16,7 @@ namespace pythonic {
             {
                 auto vind = asarray(ind);
                 auto vv = asarray(v);
-                for(size_t i=0; i<ind.flat_size(); ++i)
+                for(long i=0; i<ind.flat_size(); ++i)
                 {
                     auto val = *(vind.fbegin() + i);
                     if(val>=expr.flat_size() || val <0)
