+++ pythran/pythonic/numpy/select.hpp
@@ -53,9 +53,9 @@ namespace pythonic {
             types::ndarray<T,N> select(types::list<types::ndarray<U,N>> const& condlist, types::list<types::ndarray<T,N>> const& choicelist, T _default = 0)
             {
                 types::ndarray<T,N> out(choicelist[0].shape, _default);
-                for(size_t i=0; i<out.flat_size(); ++i)
+                for(long i=0; i<out.flat_size(); ++i)
                 {
-                    for(int j=0; j<condlist.size(); ++j)
+                    for(long j=0; j<condlist.size(); ++j)
                     {
                         if(condlist[j].buffer[i])
                         {
